Don't use size integer when it's not needed.
git-svn-id: file:///home/yoann/dev/prelude/git/nok/SVN/trunk/prelude-lml@5790 09c5ec92-17d4-0310-903a-819935f44dba

+++ plugins/pcre/pcre-mod.h
@@ -29,8 +29,8 @@
 
 
 typedef struct {
-        uint16_t id;
-        uint16_t revision;
+        unsigned int id;
+        unsigned int revision;
 
         prelude_bool_t last;
         prelude_bool_t chained;
